+++ src/__tests__/pages/teachers/programmes/[programmeSlug]/units/[unitSlug]/lessonListing.test.tsx
@@ -13,6 +13,8 @@ import lessonListingFixture, {
   lessonsWithUnpublishedContent,
 } from "@/node-lib/curriculum-api-2023/fixtures/lessonListing.fixture";
 import curriculumApi from "@/node-lib/curriculum-api-2023/__mocks__/index";
+import { setUseUserReturn } from "@/__tests__/__helpers__/mockClerk";
+import { mockLoggedIn } from "@/__tests__/__helpers__/mockUser";
 
 const render = renderWithProviders();
 
@@ -28,15 +30,24 @@ jest.mock("@/context/Analytics/useAnalytics", () => ({
   default: () => ({
     track: {
       lessonAccessed: (...args: unknown[]) => lessonSelected(...args),
+      teacherShareInitiated: () => jest.fn(),
+      contentSaved: () => jest.fn(),
+      contentUnsaved: () => jest.fn(),
     },
   }),
 }));
+
+// mock save functionality
+window.global.fetch = jest.fn().mockResolvedValue({ ok: true });
 
 jest.mock("posthog-js/react", () => ({
   useFeatureFlagEnabled: jest.fn().mockReturnValue(true),
 }));
 
 describe("Lesson listing page", () => {
+  beforeEach(() => {
+    setUseUserReturn(mockLoggedIn);
+  });
   test("it renders the unit title as page title", () => {
     const { getByRole } = render(
       <LessonListPage curriculumData={lessonListingFixture()} />,
@@ -44,6 +55,9 @@ describe("Lesson listing page", () => {
 
     const pageHeading = getByRole("heading", { level: 1 });
 
+    expect(global.fetch).toHaveBeenCalledWith(
+      "https://mockdownloads.com/api/unit/adding-surds-1/check-files",
+    );
     expect(pageHeading).toBeInTheDocument();
   });
 
@@ -54,6 +68,9 @@ describe("Lesson listing page", () => {
     const lessonCountFixtures = lessonListingFixture().lessons.length;
     const lessonCount = getByText(`Lessons (${lessonCountFixtures})`);
 
+    expect(global.fetch).toHaveBeenCalledWith(
+      "https://mockdownloads.com/api/unit/adding-surds-1/check-files",
+    );
     expect(lessonCount).toBeInTheDocument();
   });
 
@@ -85,6 +102,9 @@ describe("Lesson listing page", () => {
       const { seo } = renderWithSeo()(
         <LessonListPage curriculumData={lessonListingFixture()} />,
       );
+      expect(global.fetch).toHaveBeenCalledWith(
+        "https://mockdownloads.com/api/unit/adding-surds-1/check-files",
+      );
       expect(seo).toEqual({
         ...mockSeoResult,
         ogSiteName: "NEXT_PUBLIC_SEO_APP_NAME",
@@ -101,6 +121,9 @@ describe("Lesson listing page", () => {
       utilsMock.RESULTS_PER_PAGE = 2;
       const { seo } = renderWithSeo()(
         <LessonListPage curriculumData={lessonListingFixture()} />,
+      );
+      expect(global.fetch).toHaveBeenCalledWith(
+        "https://mockdownloads.com/api/unit/adding-surds-1/check-files",
       );
       expect(seo).toEqual({
         ...mockSeoResult,
@@ -171,9 +194,11 @@ describe("Lesson listing page", () => {
       );
 
       const lesson = getByText("Add two surds");
-
       await userEvent.click(lesson);
 
+      expect(global.fetch).toHaveBeenCalledWith(
+        "https://mockdownloads.com/api/unit/adding-surds-1/check-files",
+      );
       expect(lessonSelected).toHaveBeenCalledTimes(1);
       expect(lessonSelected).toHaveBeenCalledWith({
         platform: "owa",