+++ src/components/TeacherComponents/UnitList/UnitList.tsx
@@ -1,4 +1,4 @@
-import React, { FC, MouseEvent, useState } from "react";
+import React, { FC, MouseEvent } from "react";
 import { NextRouter, useRouter } from "next/router";
 import { useFeatureFlagEnabled } from "posthog-js/react";
 import {
@@ -31,6 +31,7 @@ import { resolveOakHref } from "@/common-lib/urls";
 import isSlugLegacy from "@/utils/slugModifiers/isSlugLegacy";
 import { PaginationProps } from "@/components/SharedComponents/Pagination/usePagination";
 import { convertSubjectToSlug } from "@/components/TeacherComponents/helpers/convertSubjectToSlug";
+import { useSaveUnits } from "@/node-lib/educator-api/helpers/saveUnits/useSaveUnits";
 
 export type Tier = {
   title: string;
@@ -185,21 +186,21 @@ const UnitList: FC<UnitListProps> = (props) => {
     isSwimming: true,
   });
 
-  const { phaseSlug, keyStageSlug, examBoardSlug } = getProgrammeFactors(props);
+  const { phaseSlug, keyStageSlug, examBoardSlug, keyStageTitle } =
+    getProgrammeFactors(props);
   const indexOfFirstLegacyUnit = units
     .map((u) => isSlugLegacy(u[0]!.programmeSlug))
     .indexOf(true);
 
-  // stub implementation of saving
+  // Saving
   const isSaveEnabled = useFeatureFlagEnabled("teacher-save-units");
-  const [savedUnitsForUser, setSavedUnitsForUser] = useState<string[]>([]);
-
-  const onSave = (unitSlug: string) => {
-    const newSavedUnits = savedUnitsForUser.includes(unitSlug)
-      ? savedUnitsForUser.filter((u) => u !== unitSlug)
-      : [...savedUnitsForUser, unitSlug];
-    setSavedUnitsForUser(newSavedUnits);
-  };
+  const { onSaveToggle, isUnitSaved } = useSaveUnits(props.programmeSlug, {
+    keyStageTitle,
+    keyStageSlug,
+    subjectTitle: props.subjectTitle,
+    subjectSlug: props.subjectSlug,
+    savedFrom: "unit_listing_save_button",
+  });
 
   const hasNewAndLegacyUnits: boolean =
     !!phaseSlug && !!newPageItems.length && !!legacyPageItems.length;
@@ -262,8 +263,8 @@ const UnitList: FC<UnitListProps> = (props) => {
               : null
           }
           optionalityUnits={getOptionalityUnits(item, onClick, router)}
-          onSave={isSaveEnabled ? onSave : undefined}
-          getIsSaved={(unitSlug) => savedUnitsForUser.includes(unitSlug)}
+          onSave={isSaveEnabled ? onSaveToggle : undefined}
+          getIsSaved={isUnitSaved}
         />
       ) : (
         item.map((unitOption) => {
@@ -310,8 +311,10 @@ const UnitList: FC<UnitListProps> = (props) => {
                 unitSlug: unitOption.slug,
                 programmeSlug: unitOption.programmeSlug,
               })}
-              onSave={isSaveEnabled ? () => onSave(unitOption.slug) : undefined}
-              isSaved={savedUnitsForUser.includes(unitOption.slug)}
+              onSave={
+                isSaveEnabled ? () => onSaveToggle(unitOption.slug) : undefined
+              }
+              isSaved={isUnitSaved(unitOption.slug)}
             />
           );
         })