+++ src/components/TeacherComponents/hooks/downloadAndShareHooks/useResourceFormState.test.tsx
@@ -342,6 +342,7 @@ describe("useResourceFormState", () => {
       );
     });
     test("should throw an error for invalid resource type", () => {
+      console.error = jest.fn();
       const invalidProps = {
         type: "invalid" as UseResourceFormStateProps["type"],
       };
@@ -351,6 +352,14 @@ describe("useResourceFormState", () => {
           useResourceFormState(invalidProps as UseResourceFormStateProps),
         ),
       ).toThrow("Invalid resource type");
+      const consoleErrorCalls = (console.error as ReturnType<typeof jest.fn>)
+        .mock.calls;
+      expect(consoleErrorCalls[0]?.[0].message).toEqual(
+        "Uncaught [Error: Invalid resource type]",
+      );
+      expect(consoleErrorCalls[1]?.[0].message).toEqual(
+        "Uncaught [Error: Invalid resource type]",
+      );
     });
   });
 });