+++ src/node-lib/curriculum-api-2023/queries/pupilUnitListing/pupilUnitListing.schema.ts
@@ -1,10 +1,10 @@
 import { z } from "zod";
-import { syntheticUnitvariantsWithLessonIdsSchema } from "@oaknational/oak-curriculum-schema";
+import { syntheticUnitvariantsWithLessonIdsByYearsSchema } from "@oaknational/oak-curriculum-schema";
 
 import { ConvertKeysToCamelCase } from "@/utils/snakeCaseConverter";
 
 export const unitBrowseDataSchema = z.array(
-  syntheticUnitvariantsWithLessonIdsSchema.omit({
+  syntheticUnitvariantsWithLessonIdsByYearsSchema.omit({
     null_unitvariant_id: true,
     base_slug: true,
   }),